% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLocation.R
\name{addLocation}
\alias{addLocation}
\title{addLocation}
\usage{
addLocation(x, fillMissing = NULL, naCellNumber = 0)
}
\arguments{
\item{x}{magpie object to be enriched with spatial information. Currently only works
for 0.5 degree data sets with 59199 or 67420 cells or a subset of it.}

\item{fillMissing}{if NULL cells missing from the total 59199 are just being ignore. If set to a value
missing cells will be added with this value (e.g. all set to 0 if fillMissing is 0)}

\item{naCellNumber}{number or string given as cell number for cells which do not have an assigned cell
number in the 59199 mapping}
}
\value{
magpie object with added spatial subdimensions
}
\description{
Function translates (if possible) cell numbers into coordinates or coordinates
into cell numbers and addes them to the spatial dimension of the object.
}
\examples{
map <- Cell2Country()
 x <- y <- population_magpie

 getCoords(x) <- map[100:109, c("lon", "lat")]
 getItems(x, "i") <- NULL
 addLocation(x)

 getCells(y) <- map$celliso[100:109]
 addLocation(y)
}
\author{
Jan Philipp Dietrich, Felicitas Beier
}
