% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell2Country.R
\name{Cell2Country}
\alias{Cell2Country}
\title{Cell2Country}
\usage{
Cell2Country()
}
\value{
data frame returning the mapping between cells and countries
}
\description{
function returning the mapping between grid and country
}
\examples{

  \dontrun{
   Cell2Country()
  }

}
\author{
Edna J. Molina Bacca
}
