% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findset.R
\name{findset}
\alias{findset}
\title{findset}
\usage{
findset(set, alias = FALSE, noset = "warning", version = NULL)
}
\arguments{
\item{set}{a MAgPIE set (e.g. "kcr"), or a vector of sets.}

\item{alias}{if TRUE, set elements are extended by '_alias'. Can be used to avoid doubling of dimnames.}

\item{noset}{if "original", a set that has no mapping is returned without changes; otherwhise a vector of length 0 is returned, plus a warning that the set does not exist.}

\item{version}{Switch between different version of the magpiesets library (use 'versionset' to load version types itself)}
}
\value{
MAgPIE set items
}
\description{
function returning set item of MAgPIE sets
}
\examples{
  \dontrun{
    findset("kcr")
  }
}
\author{
Benjamin Leon Bodirsky, Florian Humpenoeder, Kristine Karstens
}
