% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRAnames.R
\name{FRAnames}
\alias{FRAnames}
\title{FRA names}
\usage{
FRAnames(x, from = "franames", to = "magpienames", mapping = "fra_names.csv")
}
\arguments{
\item{x}{a vector of common FRA 2020 names (e.g. c("nat_reg","agb"))}

\item{from}{fra names for common magpie abbreviations, FRAnames for reverse translation}

\item{to}{Common names for official reporting names, reportincolors for typical colorcode of an item}

\item{mapping}{csv file in inst/extdata folder.}
}
\value{
vector with reporting names or volors
}
\description{
function returning the reporting names from FRA 2020 category
}
\examples{

  \dontrun{
    FRAnames("nat_reg")
    FRAnames("tece",to="reportingcolors")
  }
  
}
\author{
Abhijeet Nishra
}
