% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magpiesets-package.R
\docType{package}
\name{magpiesets-package}
\alias{magpiesets-package}
\alias{magpiesets}
\title{Magpie Sets Processing Tools}
\arguments{
\item{version}{Switch between different version of the magpiesets library
\tabular{ll}{
  default_sep18  \tab default MAgPIE 4.0 set with 49 products in 'kall'\cr
  biogas_dec18   \tab biogas extension including biogas, biochar and manure in 'kall'\cr
  biochar_jan19  \tab biochar extension including biochar and pyrolyzing\cr
}}
}
\description{
Package contains helper functions for reporting and validation (special focus on commodities)
}
\author{
Benjamin Leon Bodirsky (Maintainer: \email{bodirsky@pik-potsdam.de}), Florian Humpenoeder, 
Abhijeet Mishra, Kristine Karstens
}
