% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporthelper.R
\name{reporthelper}
\alias{reporthelper}
\title{reporthelper}
\usage{
reporthelper(
  x,
  dim = 3.1,
  level_zero_name = "All products",
  detail = TRUE,
  sort = FALSE,
  partly = FALSE,
  version = NULL
)
}
\arguments{
\item{x}{Magpie object with data that shall be reported}

\item{dim}{Dimension in which magpie products ("tece" etc) can be found}

\item{level_zero_name}{the general reporting name of the Magpie object (e.g. "Agricultural Production")}

\item{detail}{if detail=F, the subcategories of groups are not reported (e.g. "soybean" within "oilcrops")}

\item{sort}{sort items in 3rd dimension alphabetically (TRUE or FALSE)}

\item{partly}{boolean or set name, that should be reported in detail, even if it is just partly provided
within the gdx}

\item{version}{Switch between different version of the magpiesets library}
}
\value{
MAgPIE object with aggregated and renamed items in 3rd dimension
}
\description{
Aggregates MAgPIE products to standard reporting categories and changes to reporting names.
Automatically recognizes if only a reduced form of "kall" is provided.
}
\examples{
\dontrun{
x <- calories(level = "regglo", products = "kcr", attributes = "protein")
x <- reporthelper(x)
}

}
\author{
Benjamin Leon Bodirsky, Florian Humpenoeder, Kristine Karstens
}
