% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingnames.R
\name{reportingnames}
\alias{reportingnames}
\title{reportingnames}
\usage{
reportingnames(
  x,
  from = "magpienames",
  to = "reportingnames",
  mapping = "reportingnames.csv"
)
}
\arguments{
\item{x}{a vector of common magpie names (e.g. c("tece","trce"))}

\item{from}{magpienames for common magpie abbreviations, reportingnames for reverse translation}

\item{to}{reportingnames for official reporting names, reportincolors for typical colorcode of an item}

\item{mapping}{csv file in inst/extdata folder.}
}
\value{
vector with reporting names or volors
}
\description{
function returning the reporting names of a vector
}
\examples{

  \dontrun{
    reportingnames("tece")
    reportingnames("tece",to="reportingcolors")
  }
  
}
\author{
Benjamin Leon Bodirsky, Florian Humpenoeder, Patrick v. Jeetze
}
