% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingReverse.R
\name{reportingReverse}
\alias{reportingReverse}
\title{reportingReverse}
\usage{
reportingReverse(x)
}
\arguments{
\item{x}{a MAgPIE which has the format of a report file (e.g. "Demand|Agriculture|Food (Mt DM)")}
}
\value{
a MAgPIE object with higher dimensionailty
}
\description{
reverses the reporting format with tree structure (e.g. "Demand",
"Demand|Agriculture|Food (Mt DM)") into a conventional magpie format
}
\examples{
\dontrun{
budget <- calcOutput("ValidDemand")
    reportingReverse(budget)
}

}
\author{
Benjamin Leon Bodirsky
}
