% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summationhelper.R
\name{summationhelper}
\alias{summationhelper}
\title{summationhelper}
\usage{
summationhelper(x, sep = "+", dim = 3.1, check = TRUE, excludeLevels = 0)
}
\arguments{
\item{x}{MAgPIE object from validation calculations in "almost" ready format. for e.g. from x <- calcOutput("ValidDemand","FAO")}

\item{sep}{summation symbol (+,++,-)}

\item{dim}{Dimension in which the modification, should take place}

\item{check}{Switch to turn off checking routine, if FALSE. Default is TRUE.}

\item{excludeLevels}{levels to be excluded from summation, e.g. 1 for Trade in Trade|Net-Exports.}
}
\value{
MAgPIE object
}
\description{
Modifies name of variables in a magpie object to add summation symbols.
}
\details{
summationhelper should help to organize the subcategories into groups for e.g. stackplots. Notice the following hints:
\itemize{
  \item Every name should just contain one summation symbol (mostly '+').
  \item The position of the symbol (counted in '|' from left side) will determine the level.
  \item Every subitem containing the same summation symbol in the same level with the same supercategory name will be summed.
  \item Items without any summation symbol will ge ignored.
  \item Items with different summation symbols will be summed up separately.
  \item In most of the cases a summation symbol will be just placed before the last level (counted in '|' from left side).
  \item It is helpful to think about which group of items should be stacked in a stackplot.
}
  An example how a summation symbol placement could look like: 
  \preformatted{  Toplevel  
  Toplevel|+|Item 1
  Toplevel|+|Item 2
  Toplevel|Item 2|+|Subitem 1
  Toplevel|Item 2|+|Subitem 1
  Toplevel|++|Item A
  Toplevel|++|Item B
  Toplevel|Item ?}
}
\examples{
  \dontrun{
    x <- calcOutput("LanduseInitialisation",aggregate = FALSE)
    getNames(x) <- paste0("Land Cover|", reportingnames(getNames(x))," (million ha)")   
    x <- summationhelper(x)
  }

}
\author{
Abhijeet Mishra, Kristine Karstens
}
